# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TCL Stream Player API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DetectRequest {
  /** The webpage URL to scan for video sources */
  url: string;
}

/**
 * Type of video source
 */
export type VideoSourceType =
  (typeof VideoSourceType)[keyof typeof VideoSourceType];

export const VideoSourceType = {
  mp4: "mp4",
  hls: "hls",
  dash: "dash",
  webm: "webm",
  iframe: "iframe",
  unknown: "unknown",
} as const;

export interface VideoSource {
  /** Direct URL to the video source */
  url: string;
  /** Type of video source */
  type: VideoSourceType;
  /** Detected or inferred title */
  title: string;
  /** Quality label if available (e.g. 1080p, 720p) */
  quality?: string;
  /** Duration in seconds if known */
  duration?: number | null;
  /** Thumbnail URL if available */
  thumbnail?: string | null;
  /** Hostname of the source page */
  sourceHost: string;
}

export interface DetectResponse {
  videos: VideoSource[];
  /** Title of the scanned page */
  pageTitle: string;
  /** The URL that was scanned */
  scannedUrl: string;
  /** Total number of detected videos */
  count: number;
}

export interface PlaylistEntry {
  id: number;
  url: string;
  type: string;
  title: string;
  quality?: string | null;
  duration?: number | null;
  thumbnail?: string | null;
  sourceHost: string;
  addedAt: string;
}

export interface PlaylistResponse {
  entries: PlaylistEntry[];
  count: number;
}

export interface AddToPlaylistRequest {
  url: string;
  type: string;
  title: string;
  quality?: string | null;
  duration?: number | null;
  thumbnail?: string | null;
  sourceHost: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
  details?: string | null;
}