# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TCL Stream Player API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: player
    description: Video detection and playlist management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /detect:
    post:
      operationId: detectVideos
      tags: [player]
      summary: Detect videos on a webpage
      description: Fetches a URL and scans it for video sources (MP4, HLS, DASH, iframes)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectRequest"
      responses:
        "200":
          description: List of detected video sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /playlist:
    get:
      operationId: getPlaylist
      tags: [player]
      summary: Get saved playlist
      description: Returns all saved video entries
      responses:
        "200":
          description: Playlist entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistResponse"
    post:
      operationId: addToPlaylist
      tags: [player]
      summary: Add video to playlist
      description: Saves a detected video to the persistent playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToPlaylistRequest"
      responses:
        "201":
          description: Added to playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistEntry"
  /playlist/{id}:
    delete:
      operationId: removeFromPlaylist
      tags: [player]
      summary: Remove video from playlist
      description: Removes a saved video entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DetectRequest:
      type: object
      properties:
        url:
          type: string
          description: The webpage URL to scan for video sources
      required:
        - url
    VideoSource:
      type: object
      properties:
        url:
          type: string
          description: Direct URL to the video source
        type:
          type: string
          enum: [mp4, hls, dash, webm, iframe, unknown]
          description: Type of video source
        title:
          type: string
          description: Detected or inferred title
        quality:
          type: string
          description: Quality label if available (e.g. 1080p, 720p)
        duration:
          type: number
          nullable: true
          description: Duration in seconds if known
        thumbnail:
          type: string
          nullable: true
          description: Thumbnail URL if available
        sourceHost:
          type: string
          description: Hostname of the source page
      required:
        - url
        - type
        - title
        - sourceHost
    DetectResponse:
      type: object
      properties:
        videos:
          type: array
          items:
            $ref: "#/components/schemas/VideoSource"
        pageTitle:
          type: string
          description: Title of the scanned page
        scannedUrl:
          type: string
          description: The URL that was scanned
        count:
          type: integer
          description: Total number of detected videos
      required:
        - videos
        - pageTitle
        - scannedUrl
        - count
    PlaylistEntry:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        type:
          type: string
        title:
          type: string
        quality:
          type: string
          nullable: true
        duration:
          type: number
          nullable: true
        thumbnail:
          type: string
          nullable: true
        sourceHost:
          type: string
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - type
        - title
        - sourceHost
        - addedAt
    PlaylistResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistEntry"
        count:
          type: integer
      required:
        - entries
        - count
    AddToPlaylistRequest:
      type: object
      properties:
        url:
          type: string
        type:
          type: string
        title:
          type: string
        quality:
          type: string
          nullable: true
        duration:
          type: number
          nullable: true
        thumbnail:
          type: string
          nullable: true
        sourceHost:
          type: string
      required:
        - url
        - type
        - title
        - sourceHost
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true
      required:
        - error