# lib/api-zod/src/generated/types/detectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TCL Stream Player API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectRequest {
  /** The webpage URL to scan for video sources */
  url: string;
}