# lib/api-zod/src/generated/types/detectResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TCL Stream Player API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoSource } from "./videoSource";

export interface DetectResponse {
  videos: VideoSource[];
  /** Title of the scanned page */
  pageTitle: string;
  /** The URL that was scanned */
  scannedUrl: string;
  /** Total number of detected videos */
  count: number;
}