# lib/api-zod/src/generated/types/playlistEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TCL Stream Player API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistEntry {
  id: number;
  url: string;
  type: string;
  title: string;
  quality?: string | null;
  duration?: number | null;
  thumbnail?: string | null;
  sourceHost: string;
  addedAt: Date;
}