# lib/api-zod/src/generated/types/playlistResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TCL Stream Player API
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistEntry } from "./playlistEntry";

export interface PlaylistResponse {
  entries: PlaylistEntry[];
  count: number;
}