# lib/api-zod/src/generated/types/videoSource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TCL Stream Player API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoSourceType } from "./videoSourceType";

export interface VideoSource {
  /** Direct URL to the video source */
  url: string;
  /** Type of video source */
  type: VideoSourceType;
  /** Detected or inferred title */
  title: string;
  /** Quality label if available (e.g. 1080p, 720p) */
  quality?: string;
  /** Duration in seconds if known */
  duration?: number | null;
  /** Thumbnail URL if available */
  thumbnail?: string | null;
  /** Hostname of the source page */
  sourceHost: string;
}